refactor(blocks-construct): refactor the file name and code
