+++ src/features/inventory/hooks/use-inventory.ts
@@ -3,7 +3,7 @@ import { useQueryClient } from '@tanstack/react-query';
 import { useToast } from 'hooks/use-toast';
 import { useTranslation } from 'react-i18next';
 import { useErrorHandler } from 'hooks/use-error-handler';
-import { AddInventoryItemParams, UpdateInventoryItemParams } from '../types/graphql.types';
+import { AddInventoryItemParams, UpdateInventoryItemParams } from '../types/inventory.types';
 import {
   getInventory,
   addInventoryItem,
