+++ src/features/inventory/types/inventory.types.ts
